% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j_import.R
\name{neo4j_import}
\alias{neo4j_import}
\title{Upload a csv or a compressed file to Neo4J import folder - accepts .csv, .zip and .tar.gz files}
\usage{
neo4j_import(con = list(address = NULL, uid = NULL, pwd = NULL),
  source = NULL, import_dir = NULL, unzip_path = "unzip",
  gunzip_path = "gunzip", tar_path = "tar")
}
\arguments{
\item{con}{List containing three objects: address, uid, pwd as character strings providing connection to the Neo4J server}

\item{source}{Character string of local path to the csv, zip or tar.gz compressed csv file}

\item{import_dir}{Character string of path to the import directory on the Neo4J server for ssh file transfer and unzipping}

\item{unzip_path}{Path to unzip to be passed to the system command on the Neo4J server}

\item{gunzip_path}{Path to gunzip to be passed to the system command on the Neo4J server}

\item{tar_path}{Path to tar to be passed to the system command on the Neo4J server}
}
\value{
System messages confirming success or error.
}
\description{
Upload a csv or a compressed file to Neo4J import folder - accepts .csv, .zip and .tar.gz files
}
