% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j_query.R
\name{neo4j_query}
\alias{neo4j_query}
\title{Execute a query string in Neo4J using cypher-shell and capture output}
\usage{
neo4j_query(
  con = list(address = NULL, uid = NULL, pwd = NULL),
  qry = NULL,
  shell_path = "cypher-shell",
  database = NULL,
  encryption = c("default", "true", "false")
)
}
\arguments{
\item{con}{List containing three objects: bolt address, uid, pwd as character strings providing connection to the Neo4J server}

\item{qry}{Character string of the query or queries to be sent to Neo4J.  Read queries should be single queries.}

\item{shell_path}{If cypher-shell is not in the PATH system variable, the full local path to cypher-shell executable.}

\item{database}{The name of the database if working in a multi-tenency instance of Neo4J.}

\item{encryption}{Passes encryption argument to cypher-shell if necessary.}
}
\value{
A dataframe of results if the read query is successful.  A text string if an error is encountered.
Write queries will return a zero length response if successful.
If multiple read queries were submitted, only the results of the final query will be returned.
}
\description{
Execute a query string in Neo4J using cypher-shell and capture output
}
\examples{
\dontrun{
# query remote Neo4J server, with cypher-shell in local system PATH variable
con <- list(address = "bolt://bolt.my-neo4j-server.com", uid = "my_username", pwd = "my_password")
qry <- "MATCH (n) RETURN (n)"
neo4j_query(con, qry)
}

\dontrun{
# query local Neo4J Community 3.5.8 server, with cypher-shell not in local system PATH variable
graph <- list(address = "bolt://localhost:7687", uid = "neo4j", pwd = "password")
SHELL_LOC <- path.expand("~/neo4j-community-3.5.8/bin/cypher-shell")
qry <- "MATCH (n) RETURN (n)"
neo4j_query(con, qry, shell_path = SHELL_LOC)
}
}
