% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j_rmdir.R
\name{neo4j_rmdir}
\alias{neo4j_rmdir}
\title{Remove subdirectory and all its contents from the Neo4J import directory}
\usage{
neo4j_rmdir(con = list(address = NULL, uid = NULL, pwd = NULL),
  dir = NULL, import_dir = NULL)
}
\arguments{
\item{con}{List containing three objects: bolt address, uid, pwd as character strings providing connection to the Neo4J server}

\item{dir}{Character string of the import subdirectory name to be deleted on the Neo4J server}

\item{import_dir}{Character string of path to the import directory on the Neo4J server}
}
\value{
A success message if successful.  A text string if an error is encountered.
}
\description{
Remove subdirectory and all its contents from the Neo4J import directory
}
