% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j_rmfiles.R
\name{neo4j_rmfiles}
\alias{neo4j_rmfiles}
\title{Remove files from the Neo4J import directory}
\usage{
neo4j_rmfiles(con = list(address = NULL, uid = NULL, pwd = NULL),
  files = NULL, import_dir = NULL)
}
\arguments{
\item{con}{List containing three objects: address, uid, pwd as character strings providing connection to the Neo4J server}

\item{files}{Character vector of file names to be passed to the rm command on the Neo4J server.  Use files = "*" to remove all files from the import directory.}

\item{import_dir}{Character string of path to the import directory on the Neo4J server}
}
\value{
A success message if successful.  A text string if an error is encountered.
}
\description{
Remove files from the Neo4J import directory
}
