% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j_status.R
\name{neo4j_status}
\alias{neo4j_status}
\title{Check status of a local Neo4J database}
\usage{
neo4j_status(neo4j_path = "neo4j")
}
\arguments{
\item{neo4j_path}{Path to the Neo4J executable (usually in the bin directory of the Neo4J installation)}
}
\value{
System messages
}
\description{
Check status of a local Neo4J database
}
\examples{
\donttest{
# Check status local graph with neo4j executable in the system PATH variable
neo4j_status()
}
}
