% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j_stop.R
\name{neo4j_stop}
\alias{neo4j_stop}
\title{Stop a local Neo4J database}
\usage{
neo4j_stop(neo4j_path = "neo4j")
}
\arguments{
\item{neo4j_path}{Path to the Neo4J executable (usually in the bin directory of the Neo4J installation)}
}
\value{
System messages
}
\description{
Stop a local Neo4J database
}
\examples{
# if neo4j exists, stop local graph with neo4j executable in the system PATH variable
if (nzchar(Sys.which("neo4j"))) {
  neo4j_stop()
}
}
