% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j_wipe.R
\name{neo4j_wipe}
\alias{neo4j_wipe}
\title{Wipe a complete local graph database in Neo4J}
\usage{
neo4j_wipe(graph = NULL, data_path = NULL)
}
\arguments{
\item{graph}{Name of local graph database object to wiped}

\item{data_path}{Path to the local Neo4J data directory}
}
\value{
Success or error message
}
\description{
Wipe a complete local graph database in Neo4J
}
\examples{
neo4j_wipe(graph = "movies.db", data_path = path.expand("~/neo4j-community-3.5.8/data"))
}
