% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j_query.R
\name{neo4j_query}
\alias{neo4j_query}
\title{Execute a query string in Neo4J using cypher-shell and capture output}
\usage{
neo4j_query(con = list(address = NULL, uid = NULL, pwd = NULL),
  qry = NULL, shell_path = "cypher-shell")
}
\arguments{
\item{con}{List containing three objects: bolt address, uid, pwd as character strings providing connection to the Neo4J server}

\item{qry}{Character string of the query or queries to be sent to Neo4J.  Read queries should be single queries.}

\item{shell_path}{If cypher-shell is not in the PATH system variable, the full local path to cypher-shell executable
(eg `path.expand('~/neo4j-community-3.5.8/bin/cypher-shell')`).}
}
\value{
A dataframe of results if the read query is successful.  A text string if an error is encountered.
Write queries will return a zero length response if successful.
If multiple read queries were submitted, only the results of the final query will be returned.
}
\description{
Execute a query string in Neo4J using cypher-shell and capture output
}
