% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j_rmfiles.R
\name{neo4j_rmfiles}
\alias{neo4j_rmfiles}
\title{Remove files from the Neo4J import directory}
\usage{
neo4j_rmfiles(
  local = F,
  con = list(address = NULL, uid = NULL, pwd = NULL),
  files = NULL,
  import_dir = "import"
)
}
\arguments{
\item{local}{Logical indicating whether import is to a locally hosted or remotely hosted server.}

\item{con}{If remotely hosted server, list containing three objects: bolt address, uid, pwd as character strings providing connection to the Neo4J server.
uid and pwd must be for an account on the server with appropriate permissions.}

\item{files}{Character vector of file names to be removed.}

\item{import_dir}{Character string of path to the Neo4J import directory.}
}
\value{
A success message if successful.  An error message otherwise.
}
\description{
Remove files from the Neo4J import directory
}
\examples{
\dontrun{
# remove file from remote Neo4J import directory
con <- list(address = "bolt://bolt.my-neo4j-server.com", uid = "my_username", pwd = "my_password")
datafile <- "data.csv"
impdir <- "./import"
neo4j_rmfiles(con = con, files = datafile, import_dir = impdir)
}

\dontrun{
# remove file from local Neo4J Community 3.5.8 import directory
con <- list(address = "bolt://localhost:7687", uid = "neo4j", pwd = "password")
datafile <- "data.csv"
impdir <- path.expand("~/neo4j-community-3.5.8/import/")
neo4j_rmfiles(local = T, con = con, files = datafile, import_dir = impdir)
}
}
