% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo4j_wipe.R
\name{neo4j_wipe}
\alias{neo4j_wipe}
\title{Wipe a complete local graph database in Neo4J}
\usage{
neo4j_wipe(database = NULL, data_path = NULL)
}
\arguments{
\item{database}{Name of local graph database object to wipe}

\item{data_path}{Path to the local Neo4J data directory}
}
\value{
Success or error message
}
\description{
Wipe a complete local graph database in Neo4J
}
\examples{
\donttest{
# wipe movies database from local Neo4J Community 3.5.8 installation
DB_LOC <- path.expand("~/neo4j-community-3.5.8/data/")
neo4j_wipe(database = "movies", data_path = DB_LOC)
}
}
